Adjust the role names
I combined the "Reviewer" and "Coordinator" roles into one ("Logistics Coordinator") because they seem to cross paths so much. Plus once we nail down a submission form for the Speakers, many of their duties will be automated. I changed "Comms" to "Marketing" to make the name more descriptive of the duties assigned to the role. I think "Moderator" sounds nicer than "MC" and it also cuts down on the confusion by keeping "MC" and "Host" straight. Lastly, we have more than just a video editor for after the completion of the talk so I created a "Post-production Team" subgroup that contains "Editor(s)" and a "Graphics Coordinator" or "GC" for short. We'll need to cut the video but also edit the captions, create keywords, make an intro card, setup the correct metadata, etc. Two people could probably handle it, but we may want to break it out by "Video Editor" and "Caption Editor."