Adjust brownbag issue template (#125)
* Lint markdown

* Grammar and puncuation fixes

* Adjust the role names

I combined the "Reviewer" and "Coordinator" roles into one ("Logistics Coordinator") because they seem to cross paths so much. Plus once we nail down a submission form for the Speakers, many of their duties will be automated. I changed "Comms" to "Marketing" to make the name more descriptive of the duties assigned to the role. I think "Moderator" sounds nicer than "MC" and it also cuts down on the confusion by keeping "MC" and "Host" straight. Lastly, we have more than just a video editor for after the completion of the talk so I created a "Post-production Team" subgroup that contains "Editor(s)" and a "Graphics Coordinator" or "GC" for short. We'll need to cut the video but also edit the captions, create keywords, make an intro card, setup the correct metadata, etc. Two people could probably handle it, but we may want to break it out by "Video Editor" and "Caption Editor."

* Add content changes from PR #68

Enable a single issue to encompass all the content we need for the full Brownbag session lifecycle.

* Lint markdown

* Adjust wording to flow better and add links

* Remove today's event options

These options will be noted in the issue template and do not belong in the script.

* Rename file for easier hyperlinking

* Add relative links for templates and procedure

* Add further info for brownbag submission

Added a section to place the necessary info which we'll collect from the Brownbag submission form. Also added a "Data Wrangler" role because the LC got too task-heavy with transfering data from the form to the PR. The intention is to make the DW obsolete with automation.

* Format

* 'Make adjustments per Bekah's feedbac

* Update event script template

* Adjust steps 4 and 5

* Change the title convention to make file linkable

* Separate Brownbag issue from process

* removing unneeded issue templates

Co-authored-by: Dan Ott <dan@dtott.com>